Move forwardRef to own module (+18 B)
